Move Model.Stable to Model.GameObjects.ChangeTracking
